fix: move tests to a `static` bucket solution (#1589)
* fix: move tests to a `static` bucket solution

- s3 bucket is created is appropriate permissions and exposed via
  environment variable.
- serverlessrepo.amazonaws.com has access to the static s3 bucket.

* integ-tests: create s3 bucket if BYOB bucket doesnt exist

+++ tests/regression/package/regression_package_base.py
@@ -1,34 +1,38 @@
 import os
-import uuid
 import json
-import tempfile
 import time
+import tempfile
+import uuid
 from pathlib import Path
 from subprocess import Popen, PIPE
 from unittest import TestCase
 
 import boto3
 
+S3_SLEEP = 3
+
 
 class PackageRegressionBase(TestCase):
     @classmethod
     def setUpClass(cls):
         cls.region_name = os.environ.get("AWS_DEFAULT_REGION")
-        cls.bucket_name = str(uuid.uuid4())
+        cls.pre_created_bucket = os.environ.get("AWS_S3", False)
+        cls.bucket_name = cls.pre_created_bucket if cls.pre_created_bucket else str(uuid.uuid4())
         cls.test_data_path = Path(__file__).resolve().parents[2].joinpath("integration", "testdata", "package")
 
-        # Create S3 bucket
+        # Intialize S3 client
         s3 = boto3.resource("s3")
+        # Use a pre-created S3 Bucket if present else create a new one
         cls.s3_bucket = s3.Bucket(cls.bucket_name)
-        cls.s3_bucket.create()
-
-        # Given 3 seconds for all the bucket creation to complete
-        time.sleep(3)
+        if not cls.pre_created_bucket:
+            cls.s3_bucket.create()
+            time.sleep(S3_SLEEP)
 
     @classmethod
     def tearDownClass(cls):
         cls.s3_bucket.objects.all().delete()
-        cls.s3_bucket.delete()
+        if not cls.pre_created_bucket:
+            cls.s3_bucket.delete()
 
     def base_command(self, base):
         command = [base]
